Create constant variables ("types") dynamically
We should *not* assign fixed values to constant variables. If the types
actually stored in the database differ from the assigned ones, we are in
trouble. Therefor let's take database values and set them appropriate.

As we can't do this on module-level we do use Django's `ready()`
function. For more information:

https://docs.djangoproject.com/en/2.2/ref/applications/#django.apps.AppConfig.ready

Currently only concrete types which are really used in code are
provided.

Note: In reality, we don't use constants in Python. Naming them in all
capital letters is a convention to separate them from variables, however,
it does not actually prevent reassignment.