Revert migrations directories - Instead create orthos user owned links
Softlinks are created from (and others):
/usr/lib/python3.8/site-packages/orthos2/data/migrations
     -> /usr/share/orthos2/data/migrations

and must be owned by orthos user. Like that:
   cd /usr/lib/orthos2
   sudo -u orthos ./manage.py makemigrate

will dump the database model of current django source code to e.g.:
/usr/share/orthos2/data/migrations/0001_initial.py

   sudo -u orthos ./manage.py migrate

will then apply this file to the database and create table structures.

If later any modifications to the data model (database structure) is
done via source code updates, makemigrate has to be called again and
the change to the current database table structures is shows up here:
/usr/share/orthos2/data/migrations/
and has to be applied with migrate again.

Theoretically this could be done in the rpm's %post section.
But this is too dangerous and is better done manually.