Tests: Fixup login tests
The problem with the login tests was that the default of
"AUTH_ALLOW_USER_CREATION" changed. Most of the tests passed as soon as
the value for this setting was mocked to the correct value for the test.

The second problem was that the user creation site was reachable and
working if one knows the URL to it. The test correctly reported this.
The fix was to pass "AUTH_ALLOW_USER_CREATION" into the template and
disable the button in the UI. The receiving POST endpoint also was
guarded with a corresponding if statement to prevent malicious
payloads to circumvent the UI restrictions.