Admin - Display Add/Change page faster
The bug was that the Add & Change Machine page wasn't being displayed when
a high number of machines (>400) is displayed.

The root cause of the bug is that the list of Hypervisor Hosts is too long.
Since Django loads all models with their relations this takes very long when
done synchronously. Once we switch the field to asynchronous loading Django
can load the list of Hypervisors once the page is already displayed.