Netbox: Initial add of helper code
This takes part of the internal stand-alone code that was used to
synchronize Racktables and Netbox.

All code to patch and update elements in Netbox was removed, the print
statements have been converted to logger statements and Black and isort
have been applied to the code. Type hints have been added to satisfy
the static type checkers.

Co-authored-by: Enno Gotthold <enno.gotthold@suse.com>