RemotePowerTypes: Fix ForgeinKey error
This was happening due to "null=True". Since that conditation couldn't be satisfied due to a missing "default=<pk>" for the ForeignKey. The solution was to add the field with "null=True" and then set all values with dummy values. After that the field could be altered to include "null=False".